简化 neteasemusic 插件，将一些逻辑进行抽象
