+++ changelog.rst
@@ -1,8 +1,8 @@
 Changelog
 ---------
 
-2.0.0a0
-"""""""
+2.0a0
+"""""
 - 使用 fuocore 中提供的 Model
 - 尝试类似 Component 的设计（参考）
