--- feeluown/plugins/neteasemusic/requirements.txt
@@ -1,1 +1,0 @@
-pycrypto==2.6.1
