--- feeluown/plugins/neteasemusic/tests/song.json
@@ -1,155 +1,0 @@
-{
-    "equalizers": {},
-    "code": 200,
-    "songs": [
-        {
-            "album": {
-                "artist": {
-                    "id": 0,
-                    "name": "",
-                    "musicSize": 0,
-                    "briefDesc": "",
-                    "picUrl": "",
-                    "img1v1Id": 0,
-                    "alias": [],
-                    "albumSize": 0,
-                    "img1v1Url": "http://p4.music.126.net/6y-UleORITEDbvrOLV0Q8A==/5639395138885805.jpg",
-                    "picId": 0,
-                    "trans": ""
-                },
-                "blurPicUrl": "http://p3.music.126.net/SwbJDnhHO0DUDWvDXJGAfQ==/6655343883051583.jpg",
-                "companyId": 0,
-                "alias": [],
-                "company": "222 Records/Interscope Records",
-                "songs": [],
-                "size": 14,
-                "picUrl": "http://p3.music.126.net/SwbJDnhHO0DUDWvDXJGAfQ==/6655343883051583.jpg",
-                "type": "\u4e13\u8f91",
-                "copyrightId": 7003,
-                "status": 0,
-                "pic": 6655343883051583,
-                "commentThreadId": "R_AL_3_2980029",
-                "tags": "",
-                "artists": [
-                    {
-                        "id": 96266,
-                        "name": "Maroon 5",
-                        "musicSize": 0,
-                        "briefDesc": "",
-                        "picUrl": "",
-                        "img1v1Id": 0,
-                        "alias": [],
-                        "albumSize": 0,
-                        "img1v1Url": "http://p4.music.126.net/6y-UleORITEDbvrOLV0Q8A==/5639395138885805.jpg",
-                        "picId": 0,
-                        "trans": ""
-                    }
-                ],
-                "publishTime": 1409587200007,
-                "description": "",
-                "briefDesc": "",
-                "id": 2980029,
-                "picId": 6655343883051583,
-                "name": "\u2164"
-            },
-            "lMusic": {
-                "id": 50895654,
-                "volumeDelta": -0.000265076,
-                "bitrate": 96000,
-                "extension": "mp3",
-                "size": 2828237,
-                "sr": 44100,
-                "name": "Sugar",
-                "playTime": 235546,
-                "dfsId": 6636652185377132
-            },
-            "position": 5,
-            "disc": "1",
-            "rurl": null,
-            "copyFrom": "",
-            "copyrightId": 7003,
-            "bMusic": {
-                "id": 50895654,
-                "volumeDelta": -0.000265076,
-                "bitrate": 96000,
-                "extension": "mp3",
-                "size": 2828237,
-                "sr": 44100,
-                "name": "Sugar",
-                "playTime": 235546,
-                "dfsId": 6636652185377132
-            },
-            "mp3Url": "http://m2.music.126.net/XdP-TUsP9aSZtxAn1af3Pg==/6636652185377132.mp3",
-            "duration": 235546,
-            "audition": {
-                "id": 50899130,
-                "volumeDelta": 0.0,
-                "bitrate": 64000,
-                "extension": "m4a",
-                "size": 1908224,
-                "sr": 44100,
-                "name": "Sugar",
-                "playTime": 235546,
-                "dfsId": 6643249255154048
-            },
-            "crbt": null,
-            "alias": [],
-            "playedNum": 0,
-            "artists": [
-                {
-                    "id": 96266,
-                    "name": "Maroon 5",
-                    "musicSize": 0,
-                    "briefDesc": "",
-                    "picUrl": "",
-                    "img1v1Id": 0,
-                    "alias": [],
-                    "albumSize": 0,
-                    "img1v1Url": "http://p4.music.126.net/6y-UleORITEDbvrOLV0Q8A==/5639395138885805.jpg",
-                    "picId": 0,
-                    "trans": ""
-                }
-            ],
-            "mMusic": {
-                "id": 50895653,
-                "volumeDelta": 0.07,
-                "bitrate": 160000,
-                "extension": "mp3",
-                "size": 4713024,
-                "sr": 44100,
-                "name": "Sugar",
-                "playTime": 235546,
-                "dfsId": 6636652185377131
-            },
-            "no": 5,
-            "starredNum": 0,
-            "name": "Sugar",
-            "rtype": 0,
-            "popularity": 100.0,
-            "dayPlays": 0,
-            "rtUrl": null,
-            "ringtone": null,
-            "mvid": 378783,
-            "ftype": 0,
-            "status": 0,
-            "commentThreadId": "R_SO_4_29019227",
-            "starred": false,
-            "hearTime": 0,
-            "id": 29019227,
-            "score": 100,
-            "rtUrls": [],
-            "hMusic": {
-                "id": 50895652,
-                "volumeDelta": -0.33,
-                "bitrate": 320000,
-                "extension": "mp3",
-                "size": 9424990,
-                "sr": 44100,
-                "name": "Sugar",
-                "playTime": 235546,
-                "dfsId": 6636652185377130
-            },
-            "fee": 0
-        }
-    ]
-}
