+++ feeluown/plugins/neteasemusic/ui.py
@@ -9,7 +9,7 @@ from PyQt5.QtWidgets import (QHBoxLayout, QVBoxLayout, QLineEdit, QHeaderView,
                              QTableWidgetItem, QSizePolicy)
 from fuocore.models import PlaylistModel, SongModel
 
-from feeluown.widgets.components import MusicTable, LP_GroupItem, ImgLabel
+from feeluown.widgets.components import LP_GroupItem
 
 from feeluown.utils import set_alpha, parse_ms
 from feeluown.widgets.base import FLabel, FFrame, FDialog, FLineEdit, \
@@ -254,467 +254,6 @@ class _TagCellWidget(FFrame):
         self.download_tag.setFixedSize(20, 20)
 
 
-class SongsTable(MusicTable):
-    play_mv_signal = pyqtSignal([int])
-    play_song_signal = pyqtSignal([object])
-    download_song_signal = pyqtSignal([SongModel])
-    show_artist_signal = pyqtSignal([int])
-    show_album_signal = pyqtSignal([int])
-    add_song_signal = pyqtSignal([SongModel])
-    set_to_next_signal = pyqtSignal([SongModel])
-
-    def __init__(self, app, rows=0, columns=6, parent=None):
-        super().__init__(app, rows, columns, parent)
-        self._app = app
-
-        self.setObjectName('nem_songs_table')
-        self.set_theme_style()
-        self.songs = []
-
-        self.setHorizontalHeaderLabels(['', '歌曲名', '歌手', '专辑', '时长',
-                                        ''])
-        self.setColumnWidth(0, 28)
-        self.setColumnWidth(2, 150)
-        self.setColumnWidth(3, 200)
-        self.setColumnWidth(5, 100)
-        self.horizontalHeader().setSectionResizeMode(1, QHeaderView.Stretch)
-        self.setHorizontalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
-        self.setVerticalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
-
-        self.setDragEnabled(True)
-        self.setDragDropMode(QAbstractItemView.DragOnly)
-
-        self._context_menu_row = 0
-        self._drag_row = None
-        self._playlist_id = 0
-
-    @pyqtSlot()
-    def add_song_to_current_playlist(self):
-        '''do not call explicit, just a slot function'''
-        song = self.songs[self._context_menu_row]
-        self.add_song_signal.emit(song)
-
-    @pyqtSlot()
-    def set_song_to_next(self):
-        '''do not call explicit, just a slot function'''
-        song = self.songs[self._context_menu_row]
-        self.set_to_next_signal.emit(song)
-
-    @pyqtSlot()
-    def download_song(self):
-        '''do not call explicit, just a slot function'''
-        song = self.songs[self._context_menu_row]
-        self.download_song_signal.emit(song)
-
-    @pyqtSlot()
-    def remove_song_from_playlist(self):
-        '''do not call explicit, just a slot function'''
-        # FIXME
-        return
-        song = self.songs[self._context_menu_row]
-        if NPlaylistModel.del_song_from_playlist(song.mid, self._playlist_id):
-            self.removeRow(self._context_menu_row)
-            self.songs.pop(self._context_menu_row)
-            self._app.message('删除 %s 成功' % song.title)
-        else:
-            self._app.message('删除 %s 失败' % song.title, error=True)
-
-    def scroll_to_song(self, song):
-        for i, s in enumerate(self.songs):
-            if s.mid == song.mid:
-                item = self.item(i, 1)
-                self.scrollToItem(item)
-                self.setCurrentItem(item)
-                break
-
-    def set_playlist_id(self, pid):
-        self._playlist_id = pid
-
-    def is_playlist(self):
-        if self._playlist_id:
-            return True
-        return False
-
-    def add_item(self, song_model):
-        music_item = QTableWidgetItem(song_model.title)
-        # FIXME
-        album_item = QTableWidgetItem(getattr(song_model.album, 'name', 'xxx'))
-        artist_item = QTableWidgetItem(song_model.artists_name)
-        m, s = parse_ms(song_model.duration)
-        duration = QTime(0, m, s)
-        length_item = QTableWidgetItem(duration.toString('mm:ss'))
-
-        row = self.rowCount()
-        self.setRowCount(row + 1)
-        self.setItem(row, 1, music_item)
-        self.setItem(row, 2, artist_item)
-        self.setItem(row, 3, album_item)
-        self.setItem(row, 4, length_item)
-        cell_widget = _TagCellWidget(self._app)
-        # FIXME
-        #if NSongModel.local_exists(song_model):
-        #    cell_widget.set_download_tag()
-        self.setCellWidget(row, 5, cell_widget)
-
-        self.songs.append(song_model)
-
-    def _is_playlist_mine(self):
-        if self.is_playlist():
-            user = NUserModel.current_user
-            if user.is_playlist_mine(self._playlist_id):
-                return True
-        return False
-
-    def contextMenuEvent(self, event):
-        menu = QMenu()
-        add_to_current_playlist_action = QAction('添加到当前播放列表', self)
-        set_song_next_to_action = QAction('下一首播放', self)
-        download_song_action = QAction('下载该歌曲', self)
-        menu.addAction(add_to_current_playlist_action)
-        menu.addAction(set_song_next_to_action)
-        menu.addAction(download_song_action)
-
-        if self._is_playlist_mine():
-            remove_song_from_playlist_action = QAction('从歌单中删除该歌曲', self)
-            menu.addAction(remove_song_from_playlist_action)
-            remove_song_from_playlist_action.triggered.connect(
-                self.remove_song_from_playlist)
-
-        add_to_current_playlist_action.triggered.connect(
-            self.add_song_to_current_playlist)
-        set_song_next_to_action.triggered.connect(
-            self.set_song_to_next)
-        download_song_action.triggered.connect(self.download_song)
-
-        point = event.pos()
-        item = self.itemAt(point)
-        if item is not None:
-            row = self.row(item)
-            self._context_menu_row = row
-            menu.exec(event.globalPos())
-
-    @property
-    def drag_song(self):
-        if self._drag_row is not None:
-            return self.songs[self._drag_row]
-        return None
-
-    def mousePressEvent(self, event):
-        super().mousePressEvent(event)
-        point = event.pos()
-        item = self.itemAt(point)
-        if item is not None:
-            self._drag_row = self.row(item)
-
-    def on_cell_dbclick(self, row, column):
-        song = self.songs[row]
-        if column == 0:
-            # FIXME
-            #if NSongModel.mv_available(song.mvid):
-            #    self.play_mv_signal.emit(song.mvid)
-            pass
-        elif column == 1:
-            self.play_song_signal.emit(song)
-            logger.debug('Double click to play song: %s', song)
-        elif column == 2:
-            self.show_artist_signal.emit(song.artists[0].identifier)
-        elif column == 3:
-            self.show_album_signal.emit(song.album.identifier)
-
-
-class SearchBox(FLineEdit):
-    def __init__(self, app, parent=None):
-        super().__init__(parent)
-        self._app = app
-
-        self.setObjectName('search_box')
-        self.setPlaceholderText('搜索歌曲、歌手')
-        self.setToolTip('输入文字可以从当前歌单内搜索\n'
-                        '按下 Enter 将搜索网络')
-        self.set_theme_style()
-
-    def set_theme_style(self):
-        theme = self._app.theme_manager.current_theme
-        style_str = '''
-            #{0} {{
-                padding-left: 3px;
-                font-size: 14px;
-                background: transparent;
-                border: 0px;
-                border-bottom: 1px solid {1};
-                color: {2};
-                outline: none;
-            }}
-            #{0}:focus {{
-                outline: none;
-            }}
-        '''.format(self.objectName(),
-                   theme.color6.name(),
-                   theme.foreground.name())
-        self.setStyleSheet(style_str)
-
-
-class CoverImgLabel(ImgLabel):
-    def __init__(self, app, parent=None):
-        super().__init__(app, parent)
-
-        self._app = app
-        self.setFixedWidth(160)
-        self.setObjectName('n_album_img_label')
-        self.set_theme_style()
-
-
-class DescriptionLabel(FLabel):
-    def __init__(self, app, parent=None):
-        super().__init__(parent)
-
-        self._app = app
-        self.setObjectName('n_desc_container')
-        self.set_theme_style()
-        self.setSizePolicy(QSizePolicy.Preferred, QSizePolicy.Expanding)
-        self.setWordWrap(True)
-        self.setTextInteractionFlags(Qt.TextSelectableByMouse)
-
-    def set_theme_style(self):
-        theme = self._app.theme_manager.current_theme
-        style_str = '''
-            #{0} {{
-                padding-top: 5px;
-                padding-bottom: 5px;
-                background: transparent;
-                color: {1};
-            }}
-        '''.format(self.objectName(),
-                   theme.foreground.name(),
-                   theme.color0.name())
-        self.setStyleSheet(style_str)
-
-    def keyPressEvent(self, event):
-        key_code = event.key()
-        if key_code == Qt.Key_Space:
-            self._preview_dialog = DescriptionPreviewDialog(self._app)
-            preview_container = self.parent().parent()
-            self._preview_dialog.set_copy(preview_container)
-            self._preview_dialog.show()
-        else:
-            super().keyPressEvent(event)
-
-
-class DescriptionPreviewDialog(FDialog):
-    def __init__(self, app, parent=None):
-        super().__init__(parent)
-        self._app = app
-
-        self.setObjectName('n_desc_preview_dialog')
-        self.desc_container = DescriptionContainer(self._app, self)
-        self._container = FFrame(self)
-        self._container.setObjectName('n_desc_preview_dialog_container')
-        self.setWindowFlags(Qt.FramelessWindowHint)
-        self.setAttribute(Qt.WA_TranslucentBackground)
-
-        self.set_theme_style()
-        self._container_layout = QVBoxLayout(self._container)
-        self._layout = QVBoxLayout(self)
-        self.setup_ui()
-
-    def set_theme_style(self):
-        theme = self._app.theme_manager.current_theme
-        style_str = '''
-            #{0} {{
-                color: {3};
-            }}
-            #{1} {{
-                background: {2};
-                border: 5px solid {4};
-                border-radius: 5px;
-                padding: 3px;
-            }}
-        '''.format(self.objectName(),
-                   self._container.objectName(),
-                   theme.background.name(),
-                   theme.foreground.name(),
-                   theme.random_color().name())
-        self.setStyleSheet(style_str)
-
-    def setup_ui(self):
-        self._layout.setContentsMargins(0, 0, 0, 0)
-        self._layout.setSpacing(0)
-
-        self._container_layout.setContentsMargins(0, 0, 0, 0)
-        self._container_layout.setSpacing(0)
-
-        self._layout.addWidget(self._container)
-        self._container_layout.addWidget(self.desc_container)
-
-    def set_copy(self, desc_container):
-        self.desc_container.set_html(desc_container.html)
-
-    def keyPressEvent(self, event):
-        key_code = event.key()
-        if key_code == Qt.Key_Space:
-            self.close()
-        else:
-            super().keyPressEvent(event)
-
-
-class DescriptionContainer(FScrollArea):
-    def __init__(self, app, parent=None):
-        super().__init__(parent)
-
-        self._app = app
-        self.desc_label = DescriptionLabel(self._app)
-        self.setObjectName('n_desc_container')
-        self.set_theme_style()
-
-        self.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Preferred)
-        self.setHorizontalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
-        self.setVerticalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
-        self._layout = QVBoxLayout(self)
-        self.setWidget(self.desc_label)
-        self.setWidgetResizable(True)
-        self.setup_ui()
-
-    @property
-    def html(self):
-        return self.desc_label.text()
-
-    def set_theme_style(self):
-        theme = self._app.theme_manager.current_theme
-        style_str = '''
-            #{0} {{
-                border: 0;
-                background: transparent;
-            }}
-        '''.format(self.objectName(),
-                   theme.foreground.name(),
-                   theme.color0.name())
-        self.setStyleSheet(style_str)
-
-    def set_html(self, desc):
-        self.desc_label.setText(desc)
-        self.desc_label.setTextFormat(Qt.RichText)
-
-    def setup_ui(self):
-        self._layout.setContentsMargins(0, 0, 0, 0)
-        self._layout.setSpacing(0)
-
-    def keyPressEvent(self, event):
-        key_code = event.key()
-        if key_code == Qt.Key_J:
-            value = self.verticalScrollBar().value()
-            self.verticalScrollBar().setValue(value + 20)
-        elif key_code == Qt.Key_K:
-            value = self.verticalScrollBar().value()
-            self.verticalScrollBar().setValue(value - 20)
-        else:
-            super().keyPressEvent(event)
-
-
-class TableControl(FFrame):
-    def __init__(self, app, parent=None):
-        super().__init__(parent)
-        self._app = app
-
-        self.play_all_btn = FButton('▶')
-        self.search_box = SearchBox(self._app)
-        self._layout = QHBoxLayout(self)
-        self.setup_ui()
-        self.setObjectName('n_table_control')
-        self.set_theme_style()
-
-    def set_theme_style(self):
-        theme = self._app.theme_manager.current_theme
-        style_str = '''
-            QPushButton {{
-                background: transparent;
-                color: {1};
-                font-size: 16px;
-                outline: none;
-            }}
-            QPushButton:hover {{
-                color: {2};
-            }}
-        '''.format(self.objectName(),
-                   theme.foreground.name(),
-                   theme.color0.name())
-        self.setStyleSheet(style_str)
-
-    def setup_ui(self):
-        self._layout.setContentsMargins(0, 0, 0, 0)
-        self._layout.setSpacing(0)
-        self.setFixedHeight(40)
-        self.play_all_btn.setFixedSize(20, 20)
-        self.search_box.setFixedSize(160, 26)
-
-        self._layout.addSpacing(20)
-        self._layout.addWidget(self.play_all_btn)
-        self._layout.addStretch(0)
-        self._layout.addWidget(self.search_box)
-        self._layout.addSpacing(60)
-
-
-class SongsTable_Container(FFrame):
-    def __init__(self, app, parent=None):
-        super().__init__(parent)
-        self._app = app
-
-        self.songs_table = None
-        self.img_label = CoverImgLabel(self._app)
-        self.desc_container = DescriptionContainer(self._app)
-        self.info_container = FFrame(parent=self)
-        self.table_control = TableControl(self._app)
-        self._layout = QVBoxLayout(self)
-        self._info_layout = QHBoxLayout(self.info_container)
-        self.setup_ui()
-
-    def setup_ui(self):
-        self._layout.setContentsMargins(0, 0, 0, 0)
-        self._layout.setSpacing(0)
-        self._info_layout.setContentsMargins(0, 0, 0, 0)
-        self._info_layout.setSpacing(0)
-
-        self._info_layout.addWidget(self.img_label)
-        self._info_layout.addSpacing(20)
-        self._info_layout.addWidget(self.desc_container)
-
-        self._layout.addSpacing(10)
-        self._layout.addWidget(self.info_container)
-        self._layout.addSpacing(10)
-        self._layout.addWidget(self.table_control)
-
-    def set_table(self, songs_table):
-        if self.songs_table:
-            self._layout.replaceWidget(self.songs_table, songs_table)
-            self.songs_table.deleteLater()
-        else:
-            self._layout.addWidget(songs_table)
-            self._layout.addSpacing(10)
-        self.songs_table = songs_table
-
-    def load_img(self, img_url, img_name):
-        self.info_container.show()
-        event_loop = asyncio.get_event_loop()
-        future = event_loop.create_task(
-            self._app.img_ctl.get(img_url, img_name))
-        future.add_done_callback(self.set_img)
-
-    def set_img(self, future):
-        content = future.result()
-        img = QImage()
-        img.loadFromData(content)
-        pixmap = QPixmap(img)
-        if pixmap.isNull():
-            return None
-        self.img_label.setPixmap(
-            pixmap.scaledToWidth(self.img_label.width(),
-                                 mode=Qt.SmoothTransformation))
-
-    def set_desc(self, desc):
-        self.desc_container.set_html(desc)
-
-    def hide_info_container(self):
-        self.info_container.hide()
-
 
 class Ui(object):
     def __init__(self, app):
@@ -724,13 +263,8 @@ class Ui(object):
         self.login_dialog = LoginDialog(self._app, self._app)
         self.login_btn = LoginButton(self._app)
         self._lb_container = FFrame()
-        self.songs_table_container = SongsTable_Container(self._app)
-        self.fm_item = LP_GroupItem(self._app, '私人FM')
-        self.fm_item.set_img_text('Ω')
         self.recommend_item = LP_GroupItem(self._app, '每日推荐')
         self.recommend_item.set_img_text('✦')
-        self.simi_item = LP_GroupItem(self._app, '相似歌曲')
-        self.simi_item.set_img_text('∾')
 
         self._lbc_layout = QHBoxLayout(self._lb_container)
 
@@ -753,13 +287,4 @@ class Ui(object):
         if self.login_dialog.isVisible():
             self.login_dialog.hide()
         library_panel = self._app.ui.central_panel.left_panel.library_panel
-        library_panel.add_item(self.fm_item)
-        library_panel.add_item(self.simi_item)
-        self.hide_simi_item()
         library_panel.add_item(self.recommend_item)
-
-    def show_simi_item(self):
-        self.simi_item.show()
-
-    def hide_simi_item(self):
-        self.simi_item.hide()
