+++ feeluown/view.py
@@ -1,0 +1,7 @@
+from feeluown.components import SongsTable
+
+
+class PlaylistTable(SongsTable):
+    def __init__(self, app, parent=None):
+        super().__init__(parent)
+        self._app = app
